+++ public/index.html
@@ -32,6 +32,6 @@
     &lt;!&ndash; Количество фильмов &ndash;&gt;
   </section>-->
 </footer>
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
